05/18/2019
Today my group members have finished the competition which was holding by the Agriculture University of Hebei Province. This is all of the program for the competition this time, and that' s every work I have done over this 30 days. I hope it might help some people do the same work as I did.